+++ AGENTS.md
@@ -2,9 +2,8 @@
 
 ## Project Structure & Module Organization
 - Workspace root hosts `Cargo.toml` that pins resolver `2` and registers members.
-- Core library lives in `paykit-lib/` with its own `Cargo.toml` and `src/lib.rs`; treat this crate as the canonical abstraction over the routing network.
-- Transport abstractions live in `paykit-lib/src/transport/`: `traits.rs` defines the public interfaces, while feature-gated adapters (e.g., `transport/pubky/*`) provide concrete implementations.
-- `transport/pubky/mod.rs` exports `PAYKIT_PATH_PREFIX` (`/pub/paykit/v0/`) to keep all Pubky paths consistent—reuse them instead of hard-coding strings.
+- Core library lives in `paykit-lib/` with its own `Cargo.toml` and `src/lib.rs`; treat this crate as the canonical Pubky-backed Paykit implementation.
+- Pubky routing helpers live in `paykit-lib/src/pubky_routing.rs`; it owns public endpoint storage operations and exports `PAYKIT_PATH_PREFIX` (`/pub/paykit/v0/`) plus `PAYKIT_PRIVATE_PATH_PREFIX` (`/pub/paykit/v0/private`). Reuse those constants instead of hard-coding strings.
 
 ## Build, Test, and Development Commands
 - `cargo fmt` — run rustfmt on every crate; required before submitting changes.
@@ -19,24 +18,23 @@
 - Keep files ASCII; when referencing Paykit vocabulary copy spellings from `README.md`.
 - Prefer descriptive module names such as `routing`, `payments`, `endpoints` to mirror protocol sections.
 
-## Transport Abstraction & Dependency Injection
-- Keep the library stateless. Functions that touch remote state must accept `AuthenticatedTransport` or `UnauthenticatedTransportRead` implementors instead of concrete SDK types.
-- Pubky support lives behind the default `pubky` feature; adapters such as `PubkyAuthenticatedTransport` and `PubkyUnauthenticatedTransport` simply wrap `PubkySession` and `pubky::PublicStorage`. Disable the feature if you need to compile without the SDK.
-- When adding or updating adapters, follow the convention: `fetch_payment_endpoint` returns `Option`, list operations treat 404s as empty, and contact discovery relies on directory listings rather than file contents.
-- Document in each API that session creation, capability scope, and key rotation remain the caller's responsibility; Paykit only consumes the trait methods it needs.
-- Timeout handling is the transport layer's responsibility, not Paykit's. The traits do not enforce any deadline — implementations must configure their own timeouts. The Pubky SDK exposes [`PubkyHttpClientBuilder::request_timeout`](https://docs.rs/pubky/latest/pubky/struct.PubkyHttpClientBuilder.html#method.request_timeout) for this purpose.
+## Pubky Routing & Dependency Handling
+- Paykit supports Pubky as its only network/storage backend. Do not add generic transport traits, feature-gated transport adapters, or alternate-backend seams unless product direction changes explicitly.
+- Keep the library stateless. Functions that touch remote public state should accept concrete Pubky SDK handles (`pubky::PubkySession` for authenticated writes, `pubky::PublicStorage` for unauthenticated reads).
+- Public endpoint reads treat missing files/directories (404/GONE) as `None` or an empty list. Contact/payment discovery relies on directory listings rather than file contents.
+- Document in each API that session creation, capability scope, and key rotation remain the caller's responsibility; Paykit only consumes the Pubky methods it needs.
+- Timeout handling is the caller/Pubky-client responsibility, not Paykit's. Paykit does not enforce any deadline. The Pubky SDK exposes [`PubkyHttpClientBuilder::request_timeout`](https://docs.rs/pubky/latest/pubky/struct.PubkyHttpClientBuilder.html#method.request_timeout) for this purpose.
 
 ### Public vs. Private Payload Types
-- **Public** payment methods use `EndpointData` (a UTF-8 `String` wrapper) at the transport trait level. Each method is stored as a separate file at a well-known path. The transport traits (`AuthenticatedTransport`, `UnauthenticatedTransportRead`) handle public payment storage.
-- **Private** payment methods bypass the transport traits entirely. They are handled by `pubky-noise`'s `PubkyNoiseEncryptor`, which manages encryption, file naming, and storage via `send_message`/`receive_message`. Private payment plaintext is a versioned JSON envelope (`version`, `kind`, UUID-v4 `reference`, and `entries`). The `write_path` and `read_path` (asymmetric folder prefixes derived per-peer-pair via `pubky_noise::path_derivation::derive_asymmetric_paths`) are set during `initiate_encrypted_link` / `accept_encrypted_link`; pubky-noise manages individual file slots within those folders using a counter-based scheme.
+- **Public** payment methods use `EndpointData` (a UTF-8 `String` wrapper). Each method is stored as a separate file at a well-known Pubky path under `PAYKIT_PATH_PREFIX`. `set_payment_endpoint` / `remove_payment_endpoint` operate on `PubkySession`; `get_payment_list` / `get_payment_endpoint` operate on `PublicStorage`.
+- **Private** payment methods are handled by `pubky-noise`'s `PubkyNoiseEncryptor`, which manages encryption, file naming, and storage via `send_message`/`receive_message`. Private payment plaintext is a versioned JSON envelope (`version`, `kind`, UUID-v4 `reference`, and `entries`). The `write_path` and `read_path` (asymmetric folder prefixes derived per-peer-pair via `pubky_noise::path_derivation::derive_asymmetric_paths`) are set during `initiate_encrypted_link` / `accept_encrypted_link`; pubky-noise manages individual file slots within those folders using a counter-based scheme.
 - The helper functions `set_private_payments` and `get_private_payments` in `lib.rs` compose JSON serialization with `PubkyNoiseEncryptor::send_message`/`receive_message`. The caller is responsible for managing the payments map (adding/removing entries) and passing the complete map inside `PrivatePaymentsPayload` to `set_private_payments`.
 - Private payments are latest-state data: when multiple private payment envelopes are queued, `get_private_payments` returns the latest and supersedes older private-payment envelopes. Receipt access is event-like: `get_receipt_access` must return every currently available `paykit.receipt_access` message as a FIFO vector instead of using latest-wins or one-at-a-time semantics. Unsupported syntactically valid private application message kinds must be logged and dropped rather than buffered indefinitely. Future event-like kinds must preserve all messages in send order once they are explicitly recognized by Paykit.
-- Private payment helper functions are `#[cfg(feature = "pubky")]` and accept concrete Pubky types (not generic over the transport traits) because they depend on `pubky-noise` for Noise encryption.
 - The serialized private payments JSON must fit within a single pubky-noise message (`PUBKY_NOISE_MSG_LEN`, currently 1000 bytes).
 
 ## Testing Guidelines
 - Rely on the standard Rust test harness; embed minimal reproducible examples in doc comments so `cargo test` exercises them automatically.
-- Name tests using the pattern `test_<feature>_<case>()` (e.g., `test_supported_list_parsing`).
+- Name tests using the pattern `test_<feature>_<case>()` (e.g. `test_supported_list_parsing`).
 - New protocol features require at least one positive and one failure-path test.
 - Aim for full coverage on serialization/deserialization paths that map to on-chain or network data.
